Rename legacy analyze to analyze_v1
