+++ src/sparsezoo/analyze_v1/cli.py
