+++ tests/sparsezoo/analyze/test_model_analysis_creation.py
@@ -18,7 +18,7 @@ import onnx
 import pytest
 
 from sparsezoo import Model
-from sparsezoo.analyze import ModelAnalysis
+from sparsezoo.analyze_v1 import ModelAnalysis
 
 
 def onnx_stub():
